Add consulting report
I chose to write up the next steps section of report 1. Here's the draft, but I want to know your thoughts on a couple of things. Firstly, I couldn't decide whether to use an 80/20 or 70/30 data split for the training and test sets, so I went with 75/25. Second, a lot of researchers do EDA on their entire data set, but in ST 590, Dr. Post said EDA should only be done on the training data because it introduces bias (kind of like training on the entire data set does). Finally, I guess it's not imperative to include Boosted Trees in our list of models, so we can leave it out if we just want to do three, but I just thought including it couldn't hurt. Let me know your thoughts.